renamed wire_new.py to wire.py
